# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "단기 팝업스토어 턴키 예약·무인 운영 플랫폼 API. 모든 응답은 {data, error} 봉투로 감싸진다.",
    "title": "PopupReady API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "로컬 개발"
    }
  ],
...
  "paths": {
...
    "/api/v1/settlements": {
      "get": {
        "description": "결제 1건이 만든 분할 정산 Row를 모두 돌려준다. 보증금 Row는 ESCROW_HELD 상태이며 정산이 아니라 반환 대상이다.",
        "operationId": "listSettlements",
        "parameters": [
          {
            "description": "조회할 예약 요청 ID",
            "example": 1,
            "in": "query",
            "name": "reservationId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListSettlementResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 값 검증 실패"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "인증이 필요함"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "역할 또는 당사자 자격이 없음"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "서버 오류"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "분할 정산 내역 조회",
        "tags": [
          "settlement"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "에러 상세. 성공 응답에서는 null이다.",
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드 상수. 클라이언트는 이 값으로 분기한다",
            "enum": [
              "VALIDATION_FAILED",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "METHOD_NOT_ALLOWED",
              "UNSUPPORTED_MEDIA_TYPE",
              "INTERNAL_ERROR",
              "EMAIL_ALREADY_EXISTS",
              "INVALID_CREDENTIALS",
              "SPACE_NOT_FOUND",
              "FIXTURE_NOT_FOUND",
              "RESERVATION_REQUEST_NOT_FOUND",
              "LAYOUT_OUT_OF_BOUNDS",
              "LAYOUT_OVERLAP",
              "FIXTURE_STOCK_EXCEEDED",
              "FIXTURE_UNAVAILABLE",
              "POWER_LIMIT_EXCEEDED",
              "CONTRACT_NOT_FOUND",
              "NOT_CONTRACT_PARTY",
              "CONTRACT_ALREADY_EXISTS",
              "CONTRACT_ALREADY_SIGNED",
              "CONTRACT_INTEGRITY_VIOLATION",
              "REFRESH_TOKEN_INVALID",
              "PAYMENT_ALREADY_COMPLETED",
              "PAYMENT_AMOUNT_MISMATCH",
              "LOCK_ACQUISITION_FAILED",
              "DOOR_NOT_YET_OPENABLE"
            ],
            "example": "VALIDATION_FAILED"
          },
          "message": {
            "type": "string",
            "description": "사람이 읽는 설명. 분기 조건으로 쓰지 말 것",
            "example": "email은 필수입니다"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "description": "실패 응답 봉투. data는 항상 null이고 error에 코드·메시지가 담긴다.",
        "properties": {
          "data": {
            "description": "실패 시 항상 null"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "data",
          "error"
        ]
      },
...
      "ApiResponseListSettlementResponse": {
        "type": "object",
        "description": "공통 응답 봉투",
        "properties": {
          "data": {
            "type": [
              "array",
              "null"
            ],
            "description": "성공 시 페이로드. 실패 시 null",
            "items": {
              "$ref": "#/components/schemas/SettlementResponse"
            }
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ],
            "description": "실패 시 에러 상세. 성공 시 null"
          }
        },
        "required": [
          "data",
          "error"
        ]
      },
...
      "SettlementResponse": {
        "type": "object",
        "description": "분할 정산 Row",
        "properties": {
          "feeAmount": {
            "type": "integer",
            "format": "int64",
            "description": "원천 차감된 플랫폼 수수료(원)",
            "example": 630000
          },
          "grossAmount": {
            "type": "integer",
            "format": "int64",
            "description": "정산 대상 총액(원)",
            "example": 6300000
          },
          "netAmount": {
            "type": "integer",
            "format": "int64",
            "description": "실제로 이체할 금액(원)",
            "example": 5670000
          },
          "payeeId": {
            "type": "integer",
            "format": "int64",
            "description": "정산 대상 사용자 ID",
            "example": 2
          },
          "status": {
            "type": "string",
            "description": "정산 상태. 보증금만 ESCROW_HELD로 생성된다",
            "enum": [
              "PENDING",
              "ESCROW_HELD",
              "APPROVED",
              "TRANSFERRED"
            ]
          },
          "type": {
            "type": "string",
            "description": "정산 종류",
            "enum": [
              "SPACE_RENT",
              "FIXTURE_RENTAL",
              "DEPOSIT",
              "PLATFORM_FEE"
            ]
          }
        },
        "required": [
          "feeAmount",
          "grossAmount",
          "netAmount",
          "payeeId",
          "status",
          "type"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "로그인 응답의 accessToken을 Bearer로 전달",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}